pyproject.toml
[project]
name = "my-app"
version = "0.1.0"
description = ""
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "uvicorn>=0.23.2",
    "langchain-openai>=0.1.14",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "langserve[server]>=0.3.3",
    "langchain>=1.2.0",
    "duckduckgo-search>=8.1.1",
    "langchain-community>=0.4.1",
    "ddgs>=9.10.0",
]

[project.optional-dependencies]
dev = [
    "langchain-cli>=0.0.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


# ValueError: Unable to determine which files to ship inside the wheel using the following heuristics:
# https://hatch.pypa.io/latest/plugins/builder/wheel/#default-file-selection

# The most likely cause of this is that there is no directory that matches the name of your project (my_app).

# At least one file selection option must be defined in the `tool.hatch.build.targets.wheel` table, see:
# https://hatch.pypa.io/latest/config/build/

# As an example, if you intend to ship a directory named `foo` that resides within a `src` directory located at the root of     
# your project, you can define the following:

# [tool.hatch.build.targets.wheel]
# packages = ["src/foo"]
[tool.hatch.build.targets.wheel]
packages = ["my-app"]  

[dependency-groups]
dev = [
    "langchain-cli>=0.0.37",
]

[[tool.uv.index]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
